requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
filelock==3.15.4
fsspec==2024.6.1
greenlet==3.0.3
huggingface-hub==0.24.5
idna==3.7
importlib_metadata==8.2.0
Jinja2==3.1.4
llvmlite==0.41.1
MarkupSafe==2.1.5
more-itertools==10.4.0
mpmath==1.3.0
networkx==3.1
numba==0.58.1
numpy==1.24.4
openai-whisper==20231117
packaging==24.1
pillow==10.2.0
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
safetensors==0.4.4
SQLAlchemy==2.0.32
sympy==1.13.1
tiktoken==0.7.0
tokenizers==0.19.1
torch==2.4.0+cu124
torchaudio==2.4.0+cu124
torchvision==0.19.0+cu124
tqdm==4.66.5
transformers==4.44.0
typing_extensions==4.12.2
urllib3==2.2.2
zipp==3.19.2
